requirements.txt
openai==0.27.6
Flask==2.3.2
psycopg==2.9.6
python-dotenv==1.0.0